tandem-collector/requirements.txt
aiohttp
asyncpg
beautifulsoup4
loguru
pydantic
APScheduler
lxml
transformers[torch]<4.43
protobuf
sentencepiece
fasttext-wheel